+++ src/cljs/lambdacd/commons.cljs
@@ -1,5 +1,6 @@
 (ns lambdacd.commons)
 
 (defn loading-screen []
-  [:div {:key "loading-screen"}
-   [:span "Loading..."]])
+  [:div {:key "loading-screen" :class "app__loading-screen"}
+   [:span "Loading..."]
+   ])